add runtime bench
